// client/src/app/modelos/reservas.ts
export class Reservas{
    _id?: string;
    ClaveReserva?: string;
    NumSerie?: string;
    ClaveExtra?: string;
    ModeloAuto?: string;
    FechaEntrega?: Date;
    HoraEntrega?: string;
    SucursalEntrega?: string;
    FechaRegreso?: Date;
    HoraRegreso?: string;
    SucursalRegreso?: string;
    Estatus?: string;
    Total?: number;
    Correo?: string; 
}